basic sensors created and tested

+++ src/test/java/id/ac/sgu/SmartHome/TempSensorTest.java
@@ -13,8 +13,8 @@ public class TempSensorTest {
 	@Test
 	public void test() {
 		Sensor sensor = new TempSensor();
-		sensor.setValue((float) 5.1);
+		sensor.setValue(5.1);
 		System.out.println(sensor.getValue());
-		assertEquals( 5.1, (float) sensor.getValue(), 0.01);
+		assertEquals( 5.1, (double) sensor.getValue(), 0.01);
 	}
 }
