+++ src/test/java/id/ac/sgu/SmartHome/WindSensorTest.java
@@ -12,8 +12,8 @@ public class WindSensorTest {
 	@Test
 	public void test() {
 		Sensor sensor = new WindSensor();
-		sensor.setValue((float)1055.00); //1055 rpm
-		assertEquals(1055, (float) sensor.getValue(), 0.1);
+		sensor.setValue(1055.00); //1055 rpm
+		assertEquals(1055, (double) sensor.getValue(), 0.1);
 	}
 
 }
